+++ packages/react/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ai-sdk/react
 
+## 4.0.0-canary.150
+
+### Patch Changes
+
+- Updated dependencies [e3d9c0e]
+  - ai@7.0.0-canary.149
+
+## 4.0.0-canary.149
+
+### Patch Changes
+
+- Updated dependencies [2852a84]
+  - ai@7.0.0-canary.148
+
+## 4.0.0-canary.148
+
+### Patch Changes
+
+- ai@7.0.0-canary.147
+
 ## 4.0.0-canary.147
 
 ### Patch Changes