+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,26 @@
 
 ### Patch Changes
 
+- Updated dependencies [e3d9c0e]
+  - ai@7.0.0-canary.149
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
+- Updated dependencies [2852a84]
+  - ai@7.0.0-canary.148
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
+- ai@7.0.0-canary.147
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
 - ai@7.0.0-canary.146
 
 ## 0.0.1-canary.0