+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ai-sdk/svelte
 
+## 5.0.0-canary.149
+
+### Patch Changes
+
+- Updated dependencies [e3d9c0e]
+  - ai@7.0.0-canary.149
+
+## 5.0.0-canary.148
+
+### Patch Changes
+
+- Updated dependencies [2852a84]
+  - ai@7.0.0-canary.148
+
+## 5.0.0-canary.147
+
+### Patch Changes
+
+- ai@7.0.0-canary.147
+
 ## 5.0.0-canary.146
 
 ### Patch Changes